feat: Update BrandIdentity component to support position prop
This commit updates the BrandIdentity component to support a new prop called "position". The position prop can be set to either "header" or "footer" to determine the visibility of the BrandIcon and BrandName elements. This change allows for more flexibility in displaying the BrandIdentity component in different sections of the page.