refactor: Remove CORS configuration from main.ts
The code changes in this commit remove the CORS configuration from the main.ts file. This configuration was previously enabling CORS with the origin specified in the environment variable ORIGIN. However, it is no longer needed and has been removed. This simplifies the code and improves the security of the application.

Co-authored-by: Saad Khan